linera_views::backends

Module scylla_db

Implements crate::store::KeyValueStore for the ScyllaDB database.

The current connection is done via a Session and a corresponding primary key called “namespace”. The maximum number of concurrent queries is controlled by max_concurrent_queries.

Structs

ScyllaDbDatabaseInternal

Database-level connection to ScyllaDB for managing namespaces and partitions.

ScyllaDbStoreInternal

The client itself and the keeping of the count of active connections.

ScyllaDbStoreInternalConfig

The type for building a new ScyllaDB Key Value Store

Enums

ScyllaDbStoreInternalError

The error type for ScyllaDbStoreInternal

Type Aliases

ScyllaDbDatabase

The ScyllaDbDatabase composed type with metrics

ScyllaDbStoreConfig

The ScyllaDbStoreConfig input type

ScyllaDbStoreError

The combined error type for the ScyllaDbDatabase.