linera_views::backends

Module value_splitting

Adds support for large values to a given store by splitting them between several keys.

Structs

LimitedTestMemoryStore

A memory store for which the values are limited to 100 bytes and can be used for tests.

ValueSplittingDatabase

A key-value database with no size limit for values.

ValueSplittingStore

A key-value store with no size limit for values.

Enums

ValueSplittingError

The composed error type built from the inner error type.

Functions

create_value_splitting_memory_store

Provides a LimitedTestMemoryStore<()> that can be used for tests.