Struct linera_views::batch::UnorderedBatch

pub struct UnorderedBatch {
    pub key_prefix_deletions: Vec<Vec<u8>>,
    pub simple_unordered_batch: SimpleUnorderedBatch,
}

An unordered batch of deletions and insertions, together with a set of key-prefixes to delete. Key-prefix deletions must happen before the insertions and the deletions.

Fields
key_prefix_deletions: Vec<Vec<u8>>

The key-prefix deletions.

simple_unordered_batch: SimpleUnorderedBatch

The batch of deletions and insertions.

Implementations

impl UnorderedBatch

pub async fn expand_delete_prefixes<DB: DeletePrefixExpander>( self, db: &DB ) -> Result<SimpleUnorderedBatch, DB::Error>

From an UnorderedBatch, creates a SimpleUnorderedBatch that does not contain the key_prefix_deletions. This requires accessing the database to eliminate them.

pub async fn expand_colliding_prefix_deletions<DB: DeletePrefixExpander>( &mut self, db: &DB ) -> Result<(), DB::Error>

Modifies an UnorderedBatch so that the key-prefix deletions do not conflict with subsequent insertions. This may require accessing the database to compute lists of deleted keys.