Trait linera_views::batch::DeletePrefixExpander

pub trait DeletePrefixExpander {
    type Error: Debug;

    // Required method
    fn expand_delete_prefix<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key_prefix: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Vec<u8>>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

A trait to expand delete_prefix operations. Certain databases (e.g. DynamoDB) do not support the deletion by prefix. Thus we need to access the databases in order to replace a DeletePrefix by a vector of the keys to be removed.

Required Associated Types

type Error: Debug

The error type that can happen when expanding the key_prefix.

Required Methods

fn expand_delete_prefix<'life0, 'life1, 'async_trait>( &'life0 self, key_prefix: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Vec<Vec<u8>>, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Returns the list of keys to be appended to the list.