Struct linera_views::collection_view::CustomCollectionView

pub struct CustomCollectionView<C, I, W> { /* private fields */ }

A MapView that serializes the indices.

Implementations

impl<C, I, W> CustomCollectionView<C, I, W>where C: Context + Send, ViewError: From<C::Error>, I: CustomSerialize, W: View<C>,

pub async fn load_entry_mut<Q>(&mut self, index: &Q) -> Result<&mut W, ViewError>where I: Borrow<Q>, Q: CustomSerialize + ?Sized,

Loads a subview for the data at the given index in the collection. If an entry was removed before then a default entry is put on this index.

  let mut view : CustomCollectionView<_, u128, RegisterView<_,String>> = CustomCollectionView::load(context).await.unwrap();
  let subview = view.load_entry_mut(&23).await.unwrap();
  let value = subview.get();
  assert_eq!(*value, String::default());

pub async fn load_entry<Q>(&mut self, index: &Q) -> Result<&W, ViewError>where I: Borrow<Q>, Q: CustomSerialize + ?Sized,

Loads a subview for the data at the given index in the collection. If an entry was removed before then a default entry is put on this index.

  let mut view : CustomCollectionView<_, u128, RegisterView<_,String>> = CustomCollectionView::load(context).await.unwrap();
  view.load_entry_mut(&23).await.unwrap();
  let subview = view.load_entry(&23).await.unwrap();
  let value = subview.get();
  assert_eq!(*value, String::default());

pub async fn try_load_entry<Q>( &self, index: &Q ) -> Result<ReadGuardedView<'_, W>, ViewError>where I: Borrow<Q>, Q: CustomSerialize + ?Sized,

Same as load_entry_mut but for read-only access. May fail if one subview is already being visited.

  let mut view : CustomCollectionView<_, u128, RegisterView<_,String>> = CustomCollectionView::load(context).await.unwrap();
  let subview = view.try_load_entry(&23).await.unwrap();
  let value = subview.get();
  assert_eq!(*value, String::default());

pub async fn reset_entry_to_default<Q>( &mut self, index: &Q ) -> Result<(), ViewError>where I: Borrow<Q>, Q: CustomSerialize + ?Sized,

Marks the entry so that it is removed in the next flush.

  let mut view : CustomCollectionView<_, u128, RegisterView<_,String>> = CustomCollectionView::load(context).await.unwrap();
  let subview = view.load_entry_mut(&23).await.unwrap();
  let value = subview.get_mut();
  *value = String::from("Hello");
  view.reset_entry_to_default(&23).await.unwrap();
  let subview = view.load_entry_mut(&23).await.unwrap();
  let value = subview.get_mut();
  assert_eq!(*value, String::default());

pub fn remove_entry<Q>(&mut self, index: &Q) -> Result<(), ViewError>where I: Borrow<Q>, Q: CustomSerialize + ?Sized,

Removes an entry from the CollectionView. If absent nothing happens.

  let mut view : CustomCollectionView<_, u128, RegisterView<_,String>> = CustomCollectionView::load(context).await.unwrap();
  let subview = view.load_entry_mut(&23).await.unwrap();
  let value = subview.get_mut();
  assert_eq!(*value, String::default());
  view.remove_entry(&23);
  let keys = view.indices().await.unwrap();
  assert_eq!(keys.len(), 0);

pub fn extra(&self) -> &C::Extra

Gets the extra data.

impl<C, I, W> CustomCollectionView<C, I, W>where C: Context + Send, ViewError: From<C::Error>, I: Send + Debug + CustomSerialize, W: View<C> + Sync,

pub async fn indices(&self) -> Result<Vec<I>, ViewError>

Returns the list of indices in the collection in the order determined by the custom serialization.

  let mut view : CustomCollectionView<_, u128, RegisterView<_,String>> = CustomCollectionView::load(context).await.unwrap();
  view.load_entry_mut(&23).await.unwrap();
  view.load_entry_mut(&25).await.unwrap();
  let indices = view.indices().await.unwrap();
  assert_eq!(indices, vec![23, 25]);

impl<C, I, W> CustomCollectionView<C, I, W>where C: Context + Send, ViewError: From<C::Error>, I: Debug + CustomSerialize, W: View<C> + Sync,

pub async fn for_each_index_while<F>(&self, f: F) -> Result<(), ViewError>where F: FnMut(I) -> Result<bool, ViewError> + Send,

Applies a function f on each index. Indices are visited in an order determined by the custom serialization. If the function f returns false, then the loop ends prematurely.

  let mut view : CustomCollectionView<_, u128, RegisterView<_,String>> = CustomCollectionView::load(context).await.unwrap();
  view.load_entry_mut(&28).await.unwrap();
  view.load_entry_mut(&24).await.unwrap();
  view.load_entry_mut(&23).await.unwrap();
  let mut part_indices = Vec::new();
  view.for_each_index_while(|index| {
    part_indices.push(index);
    Ok(part_indices.len() < 2)
  }).await.unwrap();
  assert_eq!(part_indices, vec![23, 24]);

pub async fn for_each_index<F>(&self, f: F) -> Result<(), ViewError>where F: FnMut(I) -> Result<(), ViewError> + Send,

Applies a function on each index. Indices are visited in an order determined by the custom serialization.

  let mut view : CustomCollectionView<_, u128, RegisterView<_,String>> = CustomCollectionView::load(context).await.unwrap();
  view.load_entry_mut(&28).await.unwrap();
  view.load_entry_mut(&24).await.unwrap();
  view.load_entry_mut(&23).await.unwrap();
  let mut indices = Vec::new();
  view.for_each_index(|index| {
    indices.push(index);
    Ok(())
  }).await.unwrap();
  assert_eq!(indices, vec![23, 24, 28]);