Struct linera_views::common::ContextFromDb

pub struct ContextFromDb<E, DB> {
    pub db: DB,
    pub base_key: Vec<u8>,
    pub extra: E,
}

Implementation of the Context trait on top of a DB client implementing KeyValueStoreClient.

Fields
db: DB

The DB client, usually shared between views.

base_key: Vec<u8>

The base key for the current view.

extra: E

User-defined data attached to the view.

Implementations

impl<E, DB> ContextFromDb<E, DB>where E: Clone + Send + Sync, DB: KeyValueStoreClient + Clone + Send + Sync, DB::Error: From<Error> + Send + Sync + Error + 'static, ViewError: From<DB::Error>,

pub async fn create( db: DB, base_key: Vec<u8>, extra: E ) -> Result<Self, <ContextFromDb<E, DB> as Context>::Error>

Creates a context from db that also clears the journal before making it available.

impl<E> ContextFromDb<E, Arc<RwLock<MutexGuardArc<BTreeMap<Vec<u8, Global>, Vec<u8, Global>, Global>>>>>

pub fn new(guard: MutexGuardArc<MemoryStoreMap>, extra: E) -> Self

Creates a MemoryContext.

impl<E> ContextFromDb<E, LruCachingKeyValueClient<Arc<RwLock<MutexGuardArc<BTreeMap<Vec<u8, Global>, Vec<u8, Global>, Global>>>>>>

pub async fn new( guard: MutexGuardArc<MemoryStoreMap>, base_key: Vec<u8>, extra: E, n: usize ) -> Result<Self, ViewError>

Creates a crate::key_value_store_view::KeyValueStoreMemoryContext.

impl<E> ContextFromDb<E, ViewContainer<ContextFromDb<(), Arc<RwLock<MutexGuardArc<BTreeMap<Vec<u8, Global>, Vec<u8, Global>, Global>>>>>>>

pub async fn new( guard: MutexGuardArc<MemoryStoreMap>, base_key: Vec<u8>, extra: E ) -> Result<Self, ViewError>

Creates a KeyValueStoreMemoryContext.

impl<E: Clone + Send + Sync> ContextFromDb<E, RocksdbClient>

pub fn new(db: RocksdbClient, base_key: Vec<u8>, extra: E) -> Self

Creates a RocksdbContext.

impl<E> ContextFromDb<E, DynamoDbClient>where E: Clone + Sync + Send,

pub async fn new( table: TableName, cache_size: usize, base_key: Vec<u8>, extra: E ) -> Result<(Self, TableStatus), DynamoDbContextError>

Creates a new DynamoDbContext instance.

pub async fn from_config( config: impl Into<Config>, table: TableName, cache_size: usize, base_key: Vec<u8>, extra: E ) -> Result<(Self, TableStatus), DynamoDbContextError>

Creates a new DynamoDbContext instance from the given AWS configuration.

pub async fn with_localstack( table: TableName, cache_size: usize, base_key: Vec<u8>, extra: E ) -> Result<(Self, TableStatus), DynamoDbContextError>

Creates a new DynamoDbContext instance using a LocalStack endpoint.

Requires a LOCALSTACK_ENDPOINT environment variable with the endpoint address to connect to the LocalStack instance. Creates the table if it doesn’t exist yet, reporting a TableStatus to indicate if the table was created or if it already exists.