Struct linera_views::common::ContextFromStore

pub struct ContextFromStore<E, S> {
    pub store: S,
    pub base_key: Vec<u8>,
    pub extra: E,
}

Implementation of the Context trait on top of a DB client implementing KeyValueStore.

Fields
store: S

The DB client that is shared between views.

base_key: Vec<u8>

The base key for the current view.

extra: E

User-defined data attached to the view.

Implementations

impl<E, S> ContextFromStore<E, S>where E: Clone + Send + Sync, S: KeyValueStore + Clone + Send + Sync, S::Error: From<Error> + Send + Sync + Error + 'static, ViewError: From<S::Error>,

pub async fn create( store: S, base_key: Vec<u8>, extra: E ) -> Result<Self, <ContextFromStore<E, S> as Context>::Error>

Creates a context from store that also clears the journal before making it available.

impl<E> ContextFromStore<E, TestMemoryStore>

pub fn new(extra: E) -> Self

Creates a TestMemoryContext.

impl<E> ContextFromStore<E, TestMemoryStoreInternal>

pub fn new(extra: E) -> Self

Creates a TestMemoryContextInternal.

impl<E> ContextFromStore<E, MemoryStore>

pub fn new(max_stream_queries: usize, extra: E) -> Self

Creates a MemoryContext.

impl<E> ContextFromStore<E, LruCachingStore<MemoryStore>>

pub async fn new( base_key: Vec<u8>, extra: E, n: usize ) -> Result<Self, ViewError>

Creates a crate::key_value_store_view::KeyValueStoreMemoryContext.

impl<E> ContextFromStore<E, ViewContainer<ContextFromStore<(), MemoryStore>>>

pub async fn new(base_key: Vec<u8>, extra: E) -> Result<Self, ViewError>

Creates a KeyValueStoreMemoryContext.

impl<E: Clone + Send + Sync> ContextFromStore<E, ScyllaDbStore>

pub fn new(store: ScyllaDbStore, base_key: Vec<u8>, extra: E) -> Self

Creates a ScyllaDbContext.

impl<E: Clone + Send + Sync> ContextFromStore<E, RocksDbStore>

pub fn new(store: RocksDbStore, base_key: Vec<u8>, extra: E) -> Self

Creates a RocksDbContext.

impl<E> ContextFromStore<E, DynamoDbStore>where E: Clone + Sync + Send,

pub fn new(store: DynamoDbStore, base_key: Vec<u8>, extra: E) -> Self

Creates a new DynamoDbContext instance from the given AWS configuration.