Trait linera_views::common::AdminKeyValueStore

pub trait AdminKeyValueStore: Sized + Send {
    type Error;
    type Config: Send + Sync;

    // Required methods
    fn connect(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<Self, Self::Error>> + Send;
    fn list_all(
        config: &Self::Config,
    ) -> impl Future<Output = Result<Vec<String>, Self::Error>> + Send;
    fn exists(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<bool, Self::Error>> + Send;
    fn create(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send;
    fn delete(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send;

    // Provided methods
    fn delete_all(
        config: &Self::Config,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send { ... }
    fn maybe_create_and_connect(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<Self, Self::Error>> + Send { ... }
    fn recreate_and_connect(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<Self, Self::Error>> + Send { ... }
}

Low-level trait for the administration of stores and their namespaces.

Required Associated Types

type Error

The error type returned by the store’s methods.

type Config: Send + Sync

The configuration needed to interact with a new store.

Required Methods

fn connect( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<Self, Self::Error>> + Send

Connects to an existing namespace using the given configuration.

fn list_all( config: &Self::Config, ) -> impl Future<Output = Result<Vec<String>, Self::Error>> + Send

Obtains the list of existing namespaces.

fn exists( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<bool, Self::Error>> + Send

Tests if a given namespace exists.

fn create( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Creates a namespace. Returns an error if the namespace exists.

fn delete( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Deletes the given namespace.

Provided Methods

fn delete_all( config: &Self::Config, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Deletes all the existing namespaces.

fn maybe_create_and_connect( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<Self, Self::Error>> + Send

Initializes a storage if missing and provides it.

fn recreate_and_connect( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<Self, Self::Error>> + Send

Creates a new storage. Overwrites it if this namespace already exists.

Object Safety

This trait is not object safe.