Trait linera_views::common::CustomSerialize

pub trait CustomSerialize {
    fn to_custom_bytes<C: Context>(&self) -> Result<Vec<u8>, ViewError>
    where
        ViewError: From<<C as Context>::Error>;
    fn from_custom_bytes<C: Context>(
        short_key: &[u8]
    ) -> Result<Self, ViewError>
    where
        ViewError: From<<C as Context>::Error>,
        Self: Sized;
}

Sometimes we need a serialization that is different from the usual one and for example preserves order. The {to/from}_custom_bytes has to be coherent with the Borrow trait.

Required Methods

fn to_custom_bytes<C: Context>(&self) -> Result<Vec<u8>, ViewError>where
ViewError: From<<C as Context>::Error>,

Serializes the value

fn from_custom_bytes<C: Context>(short_key: &[u8]) -> Result<Self, ViewError>where
ViewError: From<<C as Context>::Error>,
Self: Sized,

Deserialize the vector