Trait linera_views::common::KeyValueStoreClient

pub trait KeyValueStoreClient {
    type Error: Debug;
    type Keys: KeyIterable<Self::Error>;
    type KeyValues: KeyValueIterable<Self::Error>;

    fn read_key_bytes<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn find_keys_by_prefix<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key_prefix: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Self::Keys, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn find_key_values_by_prefix<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key_prefix: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Self::KeyValues, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_batch<'life0, 'life1, 'async_trait>(
        &'life0 self,
        batch: Batch,
        base_key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn clear_journal<'life0, 'life1, 'async_trait>(
        &'life0 self,
        base_key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn read_key<'life0, 'life1, 'async_trait, V>(
        &'life0 self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Option<V>, Self::Error>> + Send + 'async_trait>>
    where
        Self::Error: From<Error>,
        V: 'async_trait + DeserializeOwned,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Low-level, asynchronous key-value operations. Useful for storage APIs not based on views.

Required Associated Types

type Error: Debug

The error type.

type Keys: KeyIterable<Self::Error>

Return type for key search operations.

type KeyValues: KeyValueIterable<Self::Error>

Return type for key-value search operations.

Required Methods

fn read_key_bytes<'life0, 'life1, 'async_trait>(
&'life0 self,
key: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Retrieves a Vec<u8> from the database using the provided key.

fn find_keys_by_prefix<'life0, 'life1, 'async_trait>(
&'life0 self,
key_prefix: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<Self::Keys, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Finds the key matching the prefix. The prefix is not included in the returned keys.

fn find_key_values_by_prefix<'life0, 'life1, 'async_trait>(
&'life0 self,
key_prefix: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<Self::KeyValues, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Finds the (key,value) pairs matching the prefix. The prefix is not included in the returned keys.

fn write_batch<'life0, 'life1, 'async_trait>(
&'life0 self,
batch: Batch,
base_key: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Writes the batch in the database with base_key the base key of the entries for the journal.

fn clear_journal<'life0, 'life1, 'async_trait>(
&'life0 self,
base_key: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Clears any journal entry that may remain. The journal is located at the base_key.

Provided Methods

fn read_key<'life0, 'life1, 'async_trait, V>(
&'life0 self,
key: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<Option<V>, Self::Error>> + Send + 'async_trait>>where
Self::Error: From<Error>,
V: 'async_trait + DeserializeOwned,
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Reads a single key and deserialize the result if present.