Trait linera_views::common::ReadableKeyValueStore

pub trait ReadableKeyValueStore<E> {
    type Keys: KeyIterable<E>;
    type KeyValues: KeyValueIterable<E>;

    const MAX_KEY_SIZE: usize;

    // Required methods
    fn max_stream_queries(&self) -> usize;
    fn read_value_bytes<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, E>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn contains_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<bool, E>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn read_multi_values_bytes<'life0, 'async_trait>(
        &'life0 self,
        keys: Vec<Vec<u8>>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Option<Vec<u8>>>, E>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn find_keys_by_prefix<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key_prefix: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Self::Keys, E>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn find_key_values_by_prefix<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key_prefix: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Self::KeyValues, E>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided methods
    fn read_value<'life0, 'life1, 'async_trait, V>(
        &'life0 self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Option<V>, E>> + Send + 'async_trait>>
       where E: From<Error>,
             V: 'async_trait + DeserializeOwned,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn read_multi_values<'life0, 'async_trait, V>(
        &'life0 self,
        keys: Vec<Vec<u8>>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Option<V>>, E>> + Send + 'async_trait>>
       where E: From<Error>,
             V: 'async_trait + DeserializeOwned + Send,
             Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Low-level, asynchronous read key-value operations. Useful for storage APIs not based on views.

Required Associated Types

type Keys: KeyIterable<E>

Returns type for key search operations.

type KeyValues: KeyValueIterable<E>

Returns type for key-value search operations.

Required Associated Constants

const MAX_KEY_SIZE: usize

The maximal size of keys that can be stored.

Required Methods

fn max_stream_queries(&self) -> usize

Retrieve the number of stream queries.

fn read_value_bytes<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, E>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieves a Vec<u8> from the database using the provided key.

fn contains_key<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<bool, E>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Test whether a key exists in the database

fn read_multi_values_bytes<'life0, 'async_trait>( &'life0 self, keys: Vec<Vec<u8>> ) -> Pin<Box<dyn Future<Output = Result<Vec<Option<Vec<u8>>>, E>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Retrieves multiple Vec<u8> from the database using the provided keys.

fn find_keys_by_prefix<'life0, 'life1, 'async_trait>( &'life0 self, key_prefix: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Self::Keys, E>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Finds the key matching the prefix. The prefix is not included in the returned keys.

fn find_key_values_by_prefix<'life0, 'life1, 'async_trait>( &'life0 self, key_prefix: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Self::KeyValues, E>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Finds the (key,value) pairs matching the prefix. The prefix is not included in the returned keys.

Provided Methods

fn read_value<'life0, 'life1, 'async_trait, V>( &'life0 self, key: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Option<V>, E>> + Send + 'async_trait>>where E: From<Error>, V: 'async_trait + DeserializeOwned, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Reads a single key and deserializes the result if present.

fn read_multi_values<'life0, 'async_trait, V>( &'life0 self, keys: Vec<Vec<u8>> ) -> Pin<Box<dyn Future<Output = Result<Vec<Option<V>>, E>> + Send + 'async_trait>>where E: From<Error>, V: 'async_trait + DeserializeOwned + Send, Self: Sync + 'async_trait, 'life0: 'async_trait,

Reads multiple keys and deserializes the results if present.

Object Safety

This trait is not object safe.