Trait linera_views::common::WritableKeyValueStore

pub trait WritableKeyValueStore<E>: Send {
    const MAX_VALUE_SIZE: usize;

    // Required methods
    fn write_batch(
        &self,
        batch: Batch,
        base_key: &[u8],
    ) -> impl Future<Output = Result<(), E>> + Send;
    fn clear_journal(
        &self,
        base_key: &[u8],
    ) -> impl Future<Output = Result<(), E>> + Send;
}

Low-level, asynchronous write key-value operations. Useful for storage APIs not based on views.

Required Associated Constants

const MAX_VALUE_SIZE: usize

The maximal size of values that can be stored.

Required Methods

fn write_batch( &self, batch: Batch, base_key: &[u8], ) -> impl Future<Output = Result<(), E>> + Send

Writes the batch in the database with base_key the base key of the entries for the journal.

fn clear_journal( &self, base_key: &[u8], ) -> impl Future<Output = Result<(), E>> + Send

Clears any journal entry that may remain. The journal is located at the base_key.

Object Safety

This trait is not object safe.