Module linera_views::dynamo_db

Helper definitions for DynamoDB storage.

Structs

DynamoDbClient

A shared DB client for DynamoDb implementing LruCaching

DynamoDbClientInternal

A DynamoDB client.

DynamoDbKeyValues

A set of key-values returned by a search query on DynamoDb.

DynamoDbKeys

A set of keys returned by a search query on DynamoDB.

TableName

A DynamoDB table name.

Enums

DynamoDbContextError

Errors that occur when using DynamoDbContext.

InvalidTableName

Error when validating a table name.

TableStatus

Status of a table at the creation time of a DynamoDbContext instance.

Constants

MAX_TRANSACT_WRITE_ITEM_SIZE

Fundamental constants in DynamoDB: The maximum size of a TransactWriteItem is 100. See https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_TransactWriteItems.html

Type Definitions

Config

The configuration to connect to DynamoDB.

DynamoDbContext

An implementation of Context based on DynamoDbClient.