Enum linera_views::dynamo_db::DynamoDbContextError

pub enum DynamoDbContextError {

    Get(Box<SdkError<GetItemError>>),
    BatchWriteItem(Box<SdkError<BatchWriteItemError>>),
    TransactWriteItem(Box<SdkError<TransactWriteItemsError>>),
    Query(Box<SdkError<QueryError>>),
    TransactUpperLimitSize,
    ZeroLengthKey,
    ZeroLengthKeyPrefix,
    DatabaseRecoveryFailed,
    ValueLengthTooLarge,
    MissingKey,
    WrongKeyType(String),
    MissingValue,
    WrongValueType(String),
    BcsError(Error),
    Endpoint(EndpointError),
    CreateTable(SdkError<CreateTableError>),

}

Errors that occur when using DynamoDbContext.

Variants

Get(Box<SdkError<GetItemError>>)

An error occurred while getting the item

BatchWriteItem(Box<SdkError<BatchWriteItemError>>)

An error occurred while writing a batch of item.

TransactWriteItem(Box<SdkError<TransactWriteItemsError>>)

An error occurred while writing a transact of items.

Query(Box<SdkError<QueryError>>)

An error occurred while doing a Query

TransactUpperLimitSize

The transact maximum size is MAX_TRANSACT_WRITE_ITEM_SIZE.

ZeroLengthKey

Key have to be of non-zero length.

ZeroLengthKeyPrefix

Key prefix have to be of non-zero length.

DatabaseRecoveryFailed

The recovery failed.

ValueLengthTooLarge

The length of the value should be at most 400KB.

MissingKey

The stored key is missing.

WrongKeyType(String)

The type of the keys was not correct (It should have been a binary blob).

MissingValue

The value attribute is missing.

WrongValueType(String)

The value was stored as the wrong type (it should be a binary blob).

BcsError(Error)

A BCS error occurred.

Endpoint(EndpointError)

An Endpoint error occurred.

CreateTable(SdkError<CreateTableError>)

An error occurred while creating the table.

Implementations

impl DynamoDbContextError

pub fn wrong_key_type(value: &AttributeValue) -> Self

Creates a DynamoDbContextError::WrongKeyType instance based on the returned value type.

Panics

If the value type is in the correct type, a binary blob.

pub fn wrong_value_type(value: &AttributeValue) -> Self

Creates a DynamoDbContextError::WrongValueType instance based on the returned value type.

Panics

If the value type is in the correct type, a binary blob.