Struct linera_views::dynamo_db::DynamoDbClient

pub struct DynamoDbClient { /* private fields */ }

A shared DB client for DynamoDb implementing LruCaching

Implementations

impl DynamoDbClient

pub async fn from_config(
config: impl Into<Config>,
table: TableName,
cache_size: usize
) -> Result<(Self, TableStatus), DynamoDbContextError>

Creation of the DynamoDbClient with an LRU caching

impl DynamoDbClient

pub async fn new(
table: TableName,
cache_size: usize
) -> Result<(Self, TableStatus), DynamoDbContextError>

Creates a new DynamoDbClientInternal instance.

pub async fn with_localstack(
table: TableName,
cache_size: usize
) -> Result<(Self, TableStatus), DynamoDbContextError>

Creates a new DynamoDbClientInternal instance using a LocalStack endpoint.

Requires a LOCALSTACK_ENDPOINT environment variable with the endpoint address to connect to the LocalStack instance. Creates the table if it doesn’t exist yet, reporting a TableStatus to indicate if the table was created or if it already exists.