Type Definition linera_views::dynamo_db::DynamoDbContext

pub type DynamoDbContext<E> = ContextFromDb<E, DynamoDbClient>;

An implementation of Context based on DynamoDbClient.

Implementations

impl<E> DynamoDbContext<E>where E: Clone + Sync + Send,

pub async fn new( table: TableName, max_concurrent_queries: Option<usize>, max_stream_queries: usize, cache_size: usize, base_key: Vec<u8>, extra: E ) -> Result<(Self, TableStatus), DynamoDbContextError>

Creates a new DynamoDbContext instance.

pub async fn from_config( config: impl Into<Config>, table: TableName, max_concurrent_queries: Option<usize>, max_stream_queries: usize, cache_size: usize, base_key: Vec<u8>, extra: E ) -> Result<(Self, TableStatus), DynamoDbContextError>

Creates a new DynamoDbContext instance from the given AWS configuration.

pub async fn with_localstack( table: TableName, max_concurrent_queries: Option<usize>, max_stream_queries: usize, cache_size: usize, base_key: Vec<u8>, extra: E ) -> Result<(Self, TableStatus), DynamoDbContextError>

Creates a new DynamoDbContext instance using a LocalStack endpoint.

Requires a LOCALSTACK_ENDPOINT environment variable with the endpoint address to connect to the LocalStack instance. Creates the table if it doesn’t exist yet, reporting a TableStatus to indicate if the table was created or if it already exists.