Struct linera_views::map_view::MapView

pub struct MapView<C, I, V> { /* private fields */ }

A View that has a type for keys. The ordering of the entries is determined by the serialization of the context.

Implementations

impl<C, I, V> MapView<C, I, V>where C: Context + Sync, ViewError: From<C::Error>, I: Serialize,

pub fn insert<Q>(&mut self, index: &Q, value: V) -> Result<(), ViewError>where I: Borrow<Q>, Q: Serialize + ?Sized,

Inserts or resets a value at an index.

  let mut map : MapView<_,u32,_> = MapView::load(context).await.unwrap();
  map.insert(&(24 as u32), String::from("Hello"));
  assert_eq!(map.get(&(24 as u32)).await.unwrap(), Some(String::from("Hello")));

pub fn remove<Q>(&mut self, index: &Q) -> Result<(), ViewError>where I: Borrow<Q>, Q: Serialize + ?Sized,

Removes a value. If absent then the operation does nothing.

  let mut map = MapView::<_,u32,String>::load(context).await.unwrap();
  map.remove(&(37 as u32));
  assert_eq!(map.get(&(37 as u32)).await.unwrap(), None);

pub fn extra(&self) -> &C::Extra

Obtains the extra data.

impl<C, I, V> MapView<C, I, V>where C: Context + Sync, ViewError: From<C::Error>, I: Serialize, V: Clone + DeserializeOwned + 'static,

pub async fn get<Q>(&self, index: &Q) -> Result<Option<V>, ViewError>where I: Borrow<Q>, Q: Serialize + ?Sized,

Reads the value at the given position, if any.

  let mut map : MapView<_, u32,_> = MapView::load(context).await.unwrap();
  map.insert(&(37 as u32), String::from("Hello"));
  assert_eq!(map.get(&(37 as u32)).await.unwrap(), Some(String::from("Hello")));
  assert_eq!(map.get(&(34 as u32)).await.unwrap(), None);

pub async fn get_mut<Q>( &mut self, index: &Q ) -> Result<Option<&mut V>, ViewError>where I: Borrow<Q>, Q: Serialize + ?Sized,

Obtains a mutable reference to a value at a given position if available

  let mut map : MapView::<_,u32,String> = MapView::load(context).await.unwrap();
  map.insert(&(37 as u32), String::from("Hello"));
  assert_eq!(map.get_mut(&(34 as u32)).await.unwrap(), None);
  let value = map.get_mut(&(37 as u32)).await.unwrap().unwrap();
  *value = String::from("Hola");
  assert_eq!(map.get(&(37 as u32)).await.unwrap(), Some(String::from("Hola")));

impl<C, I, V> MapView<C, I, V>where C: Context + Sync, ViewError: From<C::Error>, I: Sync + Send + Serialize + DeserializeOwned, V: Sync + Serialize + DeserializeOwned + 'static,

pub async fn indices(&self) -> Result<Vec<I>, ViewError>

Returns the list of indices in the map. The order is determined by serialization.

  let mut map : MapView::<_,u32,String> = MapView::load(context).await.unwrap();
  map.insert(&(37 as u32), String::from("Hello"));
  assert_eq!(map.indices().await.unwrap(), vec![37 as u32]);

pub async fn for_each_index_while<F>(&self, f: F) -> Result<(), ViewError>where F: FnMut(I) -> Result<bool, ViewError> + Send,

Applies a function f on each index. Indices are visited in an order determined by the serialization. If the function returns false, then the loop ends prematurely.

  let mut map : MapView<_, u128, String> = MapView::load(context).await.unwrap();
  map.insert(&(34 as u128), String::from("Thanks"));
  map.insert(&(37 as u128), String::from("Spasiba"));
  map.insert(&(38 as u128), String::from("Merci"));
  let mut count = 0;
  map.for_each_index_while(|_index| {
    count += 1;
    Ok(count < 2)
  }).await.unwrap();
  assert_eq!(count, 2);

pub async fn for_each_index<F>(&self, f: F) -> Result<(), ViewError>where F: FnMut(I) -> Result<(), ViewError> + Send,

Applies a function f on each index. Indices are visited in the order determined by serialization.

  let mut map : MapView<_, u128, String> = MapView::load(context).await.unwrap();
  map.insert(&(34 as u128), String::from("Hello"));
  let mut count = 0;
  map.for_each_index(|_index| {
    count += 1;
    Ok(())
  }).await.unwrap();
  assert_eq!(count, 1);

pub async fn for_each_index_value_while<F>(&self, f: F) -> Result<(), ViewError>where F: FnMut(I, V) -> Result<bool, ViewError> + Send,

Applies a function f on each index/value pair. Indices and values are visited in an order determined by serialization. If the function returns false, then the loop ends prematurely.

  let mut map : MapView<_, u128, String> = MapView::load(context).await.unwrap();
  map.insert(&(34 as u128), String::from("Thanks"));
  map.insert(&(37 as u128), String::from("Spasiba"));
  map.insert(&(38 as u128), String::from("Merci"));
  let mut values = Vec::new();
  map.for_each_index_value_while(|_index, value| {
    values.push(value);
    Ok(values.len() < 2)
  }).await.unwrap();
  assert_eq!(values.len(), 2);

pub async fn for_each_index_value<F>(&self, f: F) -> Result<(), ViewError>where F: FnMut(I, V) -> Result<(), ViewError> + Send,

Applies a function on each index/value pair. Indices and values are visited in an order determined by serialization.

  let mut map : MapView<_,Vec<u8>,_> = MapView::load(context).await.unwrap();
  map.insert(&vec![0,1], String::from("Hello"));
  let mut count = 0;
  map.for_each_index_value(|_index, _value| {
    count += 1;
    Ok(())
  }).await.unwrap();
  assert_eq!(count, 1);

impl<C, I, V> MapView<C, I, V>where C: Context + Sync, ViewError: From<C::Error>, I: Serialize, V: Default + DeserializeOwned + 'static,

pub async fn get_mut_or_default<Q>( &mut self, index: &Q ) -> Result<&mut V, ViewError>where I: Borrow<Q>, Q: Sync + Send + Serialize + ?Sized,

Obtains a mutable reference to a value at a given position. Default value if the index is missing.

  let mut map : MapView<_,u32,u128> = MapView::load(context).await.unwrap();
  let value = map.get_mut_or_default(&(34 as u32)).await.unwrap();
  assert_eq!(*value, 0 as u128);