Module linera_views::memory

Helper definitions for in-memory storage.

Enums

MemoryContextError

The error type for MemoryContext.

Functions

create_test_context

Provide a MemoryContext<()> that can be used for tests.

Type Definitions

MemoryClient

A virtual DB client where data are persisted in memory.

MemoryContext

An implementation of crate::common::Context that stores all values in memory.

MemoryStoreMap

The data is serialized in memory just like for rocksdb / dynamodb The analogue of the database is the BTreeMap