Module linera_views::reentrant_collection_view

The ReentrantCollectionView implements a map structure whose keys are ordered and the values are views with concurrent access.

Structs

ReadGuardedView

A read-only accessor for a particular subview in a ReentrantCollectionView.

ReentrantByteCollectionView

A view that supports accessing a collection of views of the same kind, indexed by Vec<u8>, possibly several subviews at a time.

ReentrantCollectionView

A view that supports accessing a collection of views of the same kind, indexed by keys, possibly several subviews at a time.

ReentrantCustomCollectionView

A view that supports accessing a collection of views of the same kind, indexed by an ordered key, possibly several subviews at a time.

WriteGuardedView

A read-write accessor for a particular subview in a ReentrantCollectionView.

Type Aliases

HashedReentrantByteCollectionView

Type wrapping ReentrantByteCollectionView while memoizing the hash.

HashedReentrantCollectionView

Type wrapping ReentrantCollectionView while memoizing the hash.

HashedReentrantCustomCollectionView

Type wrapping ReentrantCustomCollectionView while memoizing the hash.