Module linera_views::rocks_db

A storage backend for views based on RocksDB

Structs

RocksDbStore

A shared DB client for RocksDB implementing LruCaching

RocksDbStoreConfig

The initial configuration of the system

RocksDbStoreInternal

The internal client

Enums

RocksDbStoreError

The error type for RocksDbContext

Functions

create_rocks_db_common_config

Creates the common initialization for RocksDB

create_rocks_db_test_config

Returns the test config and a guard for the temporary directory

create_rocks_db_test_path

Returns the test path for RocksDB without common config.

create_rocks_db_test_store

Creates a RocksDB database client to be used for tests. The temporary directory has to be carried because if it goes out of scope then the RocksDB client can become unstable.

Type Aliases

DB

The RocksDB client that we use.

RocksDbContext

An implementation of crate::common::Context based on RocksDB