linera_views

Module store

The KeyValueStore trait and related definitions. This provides the trait definitions for the stores.

Structs

CommonStoreConfig

The common initialization parameters for the KeyValueStore

CommonStoreInternalConfig

The common initialization parameters for the KeyValueStore

Traits

AdminKeyValueStore

Low-level trait for the administration of stores and their namespaces.

KeyIterable

How to iterate over the keys returned by a search query.

KeyValueIterable

How to iterate over the key-value pairs returned by a search query.

KeyValueStore

Low-level, asynchronous write and read key-value operations. Useful for storage APIs not based on views.

KeyValueStoreError

The error type for the key-value stores.

ReadableKeyValueStore

Low-level, asynchronous read key-value operations. Useful for storage APIs not based on views.

RestrictedKeyValueStore

Low-level, asynchronous write and read key-value operations. Useful for storage APIs not based on views.

TestKeyValueStore

The functions needed for testing purposes

WithError

Define an associated KeyValueStoreError.

WritableKeyValueStore

Low-level, asynchronous write key-value operations. Useful for storage APIs not based on views.