linera_views::store

Trait KeyValueDatabase

pub trait KeyValueDatabase:
    WithError
    + Sized
    + Send
    + Sync {
    type Config: Send + Sync;
    type Store;

    // Required methods
    fn get_name() -> String;
    fn connect(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<Self, Self::Error>> + Send + Sync;
    fn open_shared(&self, root_key: &[u8]) -> Result<Self::Store, Self::Error>;
    fn open_exclusive(
        &self,
        root_key: &[u8],
    ) -> Result<Self::Store, Self::Error>;
    fn list_all(
        config: &Self::Config,
    ) -> impl Future<Output = Result<Vec<String>, Self::Error>> + Send + Sync;
    fn list_root_keys(
        &self,
    ) -> impl Future<Output = Result<Vec<Vec<u8>>, Self::Error>> + Send + Sync;
    fn exists(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<bool, Self::Error>> + Send + Sync;
    fn create(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send + Sync;
    fn delete(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send + Sync;

    // Provided methods
    fn delete_all(
        config: &Self::Config,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send + Sync { ... }
    fn maybe_create_and_connect(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<Self, Self::Error>> + Send + Sync { ... }
    fn recreate_and_connect(
        config: &Self::Config,
        namespace: &str,
    ) -> impl Future<Output = Result<Self, Self::Error>> + Send + Sync { ... }
}

The definition of a key-value database.

Required Associated Types

type Config: Send + Sync

The configuration needed to interact with a new backend.

type Store

The result of opening a partition.

Required Methods

fn get_name() -> String

The name of this database.

fn connect( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<Self, Self::Error>> + Send + Sync

Connects to an existing namespace using the given configuration.

fn open_shared(&self, root_key: &[u8]) -> Result<Self::Store, Self::Error>

Opens a shared partition starting at root_key. It is understood that the partition MAY be read and written simultaneously from other clients.

fn open_exclusive(&self, root_key: &[u8]) -> Result<Self::Store, Self::Error>

Opens an exclusive partition starting at root_key. It is assumed that the partition WILL NOT be read and written simultaneously by other clients.

IMPORTANT: This assumption is not enforced at the moment. However, future implementations may choose to return an error if another client is detected.

fn list_all( config: &Self::Config, ) -> impl Future<Output = Result<Vec<String>, Self::Error>> + Send + Sync

Obtains the list of existing namespaces.

fn list_root_keys( &self, ) -> impl Future<Output = Result<Vec<Vec<u8>>, Self::Error>> + Send + Sync

Lists the root keys of the namespace. It is possible that some root keys have no keys.

fn exists( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<bool, Self::Error>> + Send + Sync

Tests if a given namespace exists.

fn create( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<(), Self::Error>> + Send + Sync

Creates a namespace. Returns an error if the namespace exists.

fn delete( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<(), Self::Error>> + Send + Sync

Deletes the given namespace.

Provided Methods

fn delete_all( config: &Self::Config, ) -> impl Future<Output = Result<(), Self::Error>> + Send + Sync

Deletes all the existing namespaces.

fn maybe_create_and_connect( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<Self, Self::Error>> + Send + Sync

Initializes a storage if missing and provides it.

fn recreate_and_connect( config: &Self::Config, namespace: &str, ) -> impl Future<Output = Result<Self, Self::Error>> + Send + Sync

Creates a new storage. Overwrites it if this namespace already exists.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.