linera_views::store
Trait LocalReadableKeyValueStore

pub trait LocalReadableKeyValueStore: WithError {
    type Keys: KeyIterable<Self::Error>;
    type KeyValues: KeyValueIterable<Self::Error>;

    const MAX_KEY_SIZE: usize;

    // Required methods
    fn max_stream_queries(&self) -> usize;
    async fn read_value_bytes(
        &self,
        key: &[u8],
    ) -> Result<Option<Vec<u8>>, Self::Error>;
    async fn contains_key(&self, key: &[u8]) -> Result<bool, Self::Error>;
    async fn contains_keys(
        &self,
        keys: Vec<Vec<u8>>,
    ) -> Result<Vec<bool>, Self::Error>;
    async fn read_multi_values_bytes(
        &self,
        keys: Vec<Vec<u8>>,
    ) -> Result<Vec<Option<Vec<u8>>>, Self::Error>;
    async fn find_keys_by_prefix(
        &self,
        key_prefix: &[u8],
    ) -> Result<Self::Keys, Self::Error>;
    async fn find_key_values_by_prefix(
        &self,
        key_prefix: &[u8],
    ) -> Result<Self::KeyValues, Self::Error>;

    // Provided methods
    fn read_value<V: DeserializeOwned>(
        &self,
        key: &[u8],
    ) -> impl Future<Output = Result<Option<V>, Self::Error>>
       where Self: Sync { ... }
    fn read_multi_values<V: DeserializeOwned + Send>(
        &self,
        keys: Vec<Vec<u8>>,
    ) -> impl Future<Output = Result<Vec<Option<V>>, Self::Error>>
       where Self: Sync { ... }
}

Low-level, asynchronous read key-value operations. Useful for storage APIs not based on views.

Required Associated Constants

const MAX_KEY_SIZE: usize

The maximal size of keys that can be stored.

Required Associated Types

type Keys: KeyIterable<Self::Error>

Returns type for key search operations.

type KeyValues: KeyValueIterable<Self::Error>

Returns type for key-value search operations.

Required Methods

fn max_stream_queries(&self) -> usize

Retrieve the number of stream queries.

async fn read_value_bytes( &self, key: &[u8], ) -> Result<Option<Vec<u8>>, Self::Error>

Retrieves a Vec<u8> from the database using the provided key.

async fn contains_key(&self, key: &[u8]) -> Result<bool, Self::Error>

Tests whether a key exists in the database

async fn contains_keys( &self, keys: Vec<Vec<u8>>, ) -> Result<Vec<bool>, Self::Error>

Tests whether a list of keys exist in the database

async fn read_multi_values_bytes( &self, keys: Vec<Vec<u8>>, ) -> Result<Vec<Option<Vec<u8>>>, Self::Error>

Retrieves multiple Vec<u8> from the database using the provided keys.

async fn find_keys_by_prefix( &self, key_prefix: &[u8], ) -> Result<Self::Keys, Self::Error>

Finds the key matching the prefix. The prefix is not included in the returned keys.

async fn find_key_values_by_prefix( &self, key_prefix: &[u8], ) -> Result<Self::KeyValues, Self::Error>

Finds the (key,value) pairs matching the prefix. The prefix is not included in the returned keys.

Provided Methods

fn read_value<V: DeserializeOwned>( &self, key: &[u8], ) -> impl Future<Output = Result<Option<V>, Self::Error>>
where Self: Sync,

Reads a single key and deserializes the result if present.

fn read_multi_values<V: DeserializeOwned + Send>( &self, keys: Vec<Vec<u8>>, ) -> impl Future<Output = Result<Vec<Option<V>>, Self::Error>>
where Self: Sync,

Reads multiple keys and deserializes the results if present.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.