linera_views::test_utils

Function exclusive_access_admin_test

pub async fn exclusive_access_admin_test<D>(exclusive_access: bool)
where
    D: TestKeyValueDatabase,
    D::Store: KeyValueStore,

A store can be in exclusive access where it stores the absence of values or in shared access where only values are stored and (key, value) once written are never modified nor erased.

In case of no exclusive access the following scenario is checked

Store 1 deletes a key and does not mark it as missing in its cache.

Store 2 writes the key

Store 1 reads the key, but since it is not in the cache it can read it correctly.

In case of exclusive access. We have the following scenario:

Store 1 deletes a key and mark it as missing in its cache.

Store 2 writes the key (it should not be doing it)

Store 1 reads the key, see it as missing.