Function linera_views::test_utils::tombstone_triggering_test

pub async fn tombstone_triggering_test<C: KeyValueStore + Sync>(
    key_value_store: C
)

That test is especially challenging for ScyllaDB. In its default settings, Scylla has a limitation to 10000 tombstones. A tombstone is an indication that the data has been deleted. That is thus a trie data structure for checking whether a requested key is deleted or not.

In this test we insert 200000 keys into the database. Then we select half of them at random and delete them. By the random selection, Scylla is forced to introduce around 100000 tombstones which triggers the crash with the default settings.