Module linera_views::value_splitting

The code for handling big values by splitting them into several small ones.

Structs

TestMemoryStore

Supposed to be removed later

TestMemoryStoreInternal

A virtual DB store where data are persisted in memory.

ValueSplittingStore

A key-value store with no size limit for values.

Enums

DatabaseConsistencyError

Data type indicating that the database is not consistent

Functions

create_test_memory_context

Provides a TestMemoryContext<()> that can be used for tests. It is not named create_memory_test_context because it is massively used and so we want to have a short name.

create_test_memory_context_internal

Provides a TestMemoryContextInternal<()> that can be used for tests.

create_test_memory_store

Creates a TestMemoryStore for working.

create_test_memory_store_internal

Provides a TestMemoryStoreInternal<()> that can be used for tests.

Type Aliases

TestMemoryContext

An implementation of crate::common::Context that stores all values in memory.

TestMemoryContextInternal

An implementation of crate::common::Context that stores all values in memory.