linera_views::views

Module collection_view

The CollectionView implements a map structure whose keys are ordered and the values are views.

Structs

ByteCollectionView

A view that supports accessing a collection of views of the same kind, indexed by a Vec<u8>, one subview at a time.

CollectionView

A view that supports accessing a collection of views of the same kind, indexed by a key, one subview at a time.

CustomCollectionView

A map view that serializes the indices.

Enums

ReadGuardedView

A read-only accessor for a particular subview in a CollectionView.

Type Aliases

HashedByteCollectionView

Type wrapping ByteCollectionView while memoizing the hash.

HashedCollectionView

Type wrapping CollectionView while memoizing the hash.

HashedCustomCollectionView

Type wrapping CustomCollectionView while memoizing the hash.

HistoricallyHashedByteCollectionView

Wrapper around ByteCollectionView to compute hashes based on the history of changes.

HistoricallyHashedCollectionView

Wrapper around CollectionView to compute hashes based on the history of changes.

HistoricallyHashedCustomCollectionView

Wrapper around CustomCollectionView to compute hashes based on the history of changes.