Struct linera_views::views::reentrant_collection_view::ReentrantCustomCollectionView

pub struct ReentrantCustomCollectionView<C, I, W> { /* private fields */ }

A view that supports accessing a collection of views of the same kind, indexed by an ordered key, possibly several subviews at a time.

Implementations

impl<C, I, W> ReentrantCustomCollectionView<C, I, W>
where C: Context + Send, ViewError: From<C::Error>, I: Sync + Clone + Send + CustomSerialize, W: View<C> + Send + Sync,

pub async fn try_load_entry_mut<Q>( &mut self, index: &Q, ) -> Result<WriteGuardedView<W>, ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Loads a subview for the data at the given index in the collection. If an entry is absent then a default entry is put in the collection on this index.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
let subview = view.try_load_entry_mut(&23).await.unwrap();
let value = subview.get();
assert_eq!(*value, String::default());

pub async fn try_load_entry<Q>( &self, index: &Q, ) -> Result<Option<ReadGuardedView<W>>, ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Loads a subview at the given index in the collection and gives read-only access to the data. If an entry is absent then None is returned.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
{
    let _subview = view.try_load_entry_mut(&23).await.unwrap();
}
let subview = view.try_load_entry(&23).await.unwrap().unwrap();
let value = subview.get();
assert_eq!(*value, String::default());

pub async fn contains_key<Q>(&self, index: &Q) -> Result<bool, ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Returns true if the collection contains a value for the specified key.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
let _subview = view.try_load_entry_mut(&23).await.unwrap();
assert!(view.contains_key(&23).await.unwrap());
assert!(!view.contains_key(&24).await.unwrap());

pub fn remove_entry<Q>(&mut self, index: &Q) -> Result<(), ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Removes an entry. If absent then nothing happens.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
let mut subview = view.try_load_entry_mut(&23).await.unwrap();
let value = subview.get_mut();
assert_eq!(*value, String::default());
view.remove_entry(&23);
let keys = view.indices().await.unwrap();
assert_eq!(keys.len(), 0);

pub fn try_reset_entry_to_default<Q>( &mut self, index: &Q, ) -> Result<(), ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Marks the entry so that it is removed in the next flush.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
{
    let mut subview = view.try_load_entry_mut(&23).await.unwrap();
    let value = subview.get_mut();
    *value = String::from("Hello");
}
{
    view.try_reset_entry_to_default(&23).unwrap();
    let subview = view.try_load_entry(&23).await.unwrap().unwrap();
    let value = subview.get();
    assert_eq!(*value, String::default());
}

pub fn extra(&self) -> &C::Extra

Gets the extra data.

impl<C, I, W> ReentrantCustomCollectionView<C, I, W>
where C: Context + Send + Clone + 'static, ViewError: From<C::Error>, I: Sync + Clone + Send + CustomSerialize, W: View<C> + Send + Sync + 'static,

pub async fn try_load_entries_mut<Q>( &mut self, indices: impl IntoIterator<Item = Q>, ) -> Result<Vec<WriteGuardedView<W>>, ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Load multiple entries for writing at once. The entries in indices have to be all distinct.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
let indices = vec![23, 42];
let subviews = view.try_load_entries_mut(indices).await.unwrap();
let value1 = subviews[0].get();
let value2 = subviews[1].get();
assert_eq!(*value1, String::default());
assert_eq!(*value2, String::default());

pub async fn try_load_entries<Q>( &self, indices: impl IntoIterator<Item = Q>, ) -> Result<Vec<Option<ReadGuardedView<W>>>, ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Load multiple entries for reading at once. The entries in indices have to be all distinct.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
{
    let _subview = view.try_load_entry_mut(&23).await.unwrap();
}
let indices = vec![23, 42];
let subviews = view.try_load_entries(indices).await.unwrap();
assert!(subviews[1].is_none());
let value0 = subviews[0].as_ref().unwrap().get();
assert_eq!(*value0, String::default());

pub async fn try_load_all_entries_mut<Q>( &mut self, ) -> Result<Vec<(I, WriteGuardedView<W>)>, ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Loads all entries for writing at once. The entries in indices have to be all distinct.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
{
    let _subview = view.try_load_entry_mut(&23).await.unwrap();
}
let subviews = view.try_load_all_entries_mut().await.unwrap();
assert_eq!(subviews.len(), 1);

pub async fn try_load_all_entries<Q>( &self, ) -> Result<Vec<(I, ReadGuardedView<W>)>, ViewError>
where I: Borrow<Q>, Q: CustomSerialize,

Load multiple entries for reading at once. The entries in indices have to be all distinct.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
{
    let _subview = view.try_load_entry_mut(&23).await.unwrap();
}
let subviews = view.try_load_all_entries().await.unwrap();
assert_eq!(subviews.len(), 1);

impl<C, I, W> ReentrantCustomCollectionView<C, I, W>
where C: Context + Send, ViewError: From<C::Error>, I: Sync + Clone + Send + CustomSerialize, W: View<C> + Send + Sync,

pub async fn indices(&self) -> Result<Vec<I>, ViewError>

Returns the list of indices in the collection. The order is determined by the custom serialization.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
view.try_load_entry_mut(&23).await.unwrap();
view.try_load_entry_mut(&25).await.unwrap();
let indices = view.indices().await.unwrap();
assert_eq!(indices, vec![23, 25]);

pub async fn count(&self) -> Result<usize, ViewError>

Returns the number of entries in the collection.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
view.try_load_entry_mut(&23).await.unwrap();
view.try_load_entry_mut(&25).await.unwrap();
assert_eq!(view.count().await.unwrap(), 2);

pub async fn for_each_index_while<F>(&self, f: F) -> Result<(), ViewError>
where F: FnMut(I) -> Result<bool, ViewError> + Send,

Applies a function f on each index. Indices are visited in an order determined by the custom serialization. If the function f returns false then the loop ends prematurely.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
view.try_load_entry_mut(&28).await.unwrap();
view.try_load_entry_mut(&24).await.unwrap();
view.try_load_entry_mut(&23).await.unwrap();
let mut part_indices = Vec::new();
view.for_each_index_while(|index| {
    part_indices.push(index);
    Ok(part_indices.len() < 2)
})
.await
.unwrap();
assert_eq!(part_indices, vec![23, 24]);

pub async fn for_each_index<F>(&self, f: F) -> Result<(), ViewError>
where F: FnMut(I) -> Result<(), ViewError> + Send,

Applies a function f on each index. Indices are visited in an order determined by the custom serialization.

let mut view: ReentrantCustomCollectionView<_, u128, RegisterView<_, String>> =
    ReentrantCustomCollectionView::load(context).await.unwrap();
view.try_load_entry_mut(&28).await.unwrap();
view.try_load_entry_mut(&24).await.unwrap();
view.try_load_entry_mut(&23).await.unwrap();
let mut indices = Vec::new();
view.for_each_index(|index| {
    indices.push(index);
    Ok(())
})
.await
.unwrap();
assert_eq!(indices, vec![23, 24, 28]);