linera_views::views

Module set_view

The SetView implements a set with ordered entries.

Structs

ByteSetView

A View that supports inserting and removing values indexed by a key.

CustomSetView

A View implementing the set functionality with the index I being a type with a custom serialization format.

SetView

A View implementing the set functionality with the index I being any serializable type.

Type Aliases

HashedByteSetView

Type wrapping ByteSetView while memoizing the hash.

HashedCustomSetView

Type wrapping CustomSetView while memoizing the hash.

HashedSetView

Type wrapping SetView while memoizing the hash.

HistoricallyHashedByteSetView

Wrapper around ByteSetView to compute hashes based on the history of changes.

HistoricallyHashedCustomSetView

Wrapper around CustomSetView to compute hashes based on the history of changes.

HistoricallyHashedSetView

Wrapper around SetView to compute hashes based on the history of changes.