Trait linera_views::views::HashableView

pub trait HashableView<C>: View<C> {
    type Hasher: Hasher;

    fn hash_mut<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<<Self::Hasher as Hasher>::Output, ViewError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn hash<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<<Self::Hasher as Hasher>::Output, ViewError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

A view that supports hashing its values.

Required Associated Types

type Hasher: Hasher

How to compute hashes.

Required Methods

fn hash_mut<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<<Self::Hasher as Hasher>::Output, ViewError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Compute the hash of the values.

Implementations do not need to include a type tag. However, the usual precautions to enforce collision-resistance must be applied (e.g. including the length of a collection of values).

fn hash<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<<Self::Hasher as Hasher>::Output, ViewError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Compute the hash of the values.

Implementations do not need to include a type tag. However, the usual precautions to enforce collision-resistance must be applied (e.g. including the length of a collection of values).