Trait linera_views::views::View

pub trait View<C>: Sized {
    fn context(&self) -> &C;
    fn load<'async_trait>(
        context: C
    ) -> Pin<Box<dyn Future<Output = Result<Self, ViewError>> + Send + 'async_trait>>
    where
        Self: 'async_trait;
    fn rollback(&mut self);
    fn clear(&mut self);
    fn flush(&mut self, batch: &mut Batch) -> Result<(), ViewError>;
    fn delete(self, batch: &mut Batch);
}

A view gives an exclusive access to read and write the data stored at an underlying address in storage.

Required Methods

fn context(&self) -> &C

Obtain a mutable reference to the internal context.

fn load<'async_trait>(
context: C
) -> Pin<Box<dyn Future<Output = Result<Self, ViewError>> + Send + 'async_trait>>where
Self: 'async_trait,

Creates a view or a subview.

fn rollback(&mut self)

Discard all pending changes. After that flush should have no effect to storage.

fn clear(&mut self)

Clears the view. That can be seen as resetting to default. In the case of a RegisterView this means setting the value to T::default(). For LogView, QueueView, this leaves the range data to be left in the database.

fn flush(&mut self, batch: &mut Batch) -> Result<(), ViewError>

Persist changes to storage. This leaves the view still usable and is essentially neutral to the program running. Crash-resistant storage implementations are expected to accumulate the desired changes in the batch variable first. If the view is dropped without calling flush, staged changes are simply lost.

fn delete(self, batch: &mut Batch)

Instead of persisting changes, clear all the data that belong to this view and its subviews. Crash-resistant storage implementations are expected to accumulate the desired changes into the batch variable first. No data/metadata at all is left after delete. The view is consumed by delete and cannot be used in any way after delete.