Crate linera_views

This module aims to help the mapping of complex data-structures onto a key-value store. The central notion is a views::View which can be loaded from storage, modified in memory, then committed (i.e. the changes are atomically persisted in storage).

Modules

batch

The definition of the batches for writing in the datatbase.

collection_view

The collection_view

common

The definitions used for the memory/rocksdb/dynamo_db

dynamo_db

Helper definitions for DynamoDB storage.

hashable_wrapper

Wrapping a view to compute a hash

key_value_store_view

The key value store view

localstack

Helper types for interfacing with a LocalStack instance.

log_view

The log_view

map_view

The map_view

memory

Helper definitions for in-memory storage.

queue_view

The queue_view

reentrant_collection_view

The reentrant_collection_view

register_view

The register_view

rocksdb

Helper definitions for Rocksdb storage.

set_view

The map_view

test_utils

Helper types for tests.

views

The main definitions.

Constants

LOAD_VIEW_COUNTER

The metric counting how often a view is read from storage.

SAVE_VIEW_COUNTER

The metric counting how often a view is written from storage.

Functions

increment_counter

Increments the metrics counter with the given name, with the struct and base key as labels.