linera_views::backends

Module rocks_db

Implements crate::store::KeyValueStore for the RocksDB database.

Structs

PathWithGuard

A path and the guard for the temporary directory if needed

RocksDbDatabaseInternal

Database-level connection to RocksDB for managing namespaces and partitions.

RocksDbStoreInternal

The inner client

RocksDbStoreInternalConfig

The initial configuration of the system

Enums

RocksDbSpawnMode

The choice of the spawning mode. SpawnBlocking always works and is the safest. BlockInPlace can only be used in multi-threaded environment. One way to select that is to select BlockInPlace when tokio::runtime::Handle::current().metrics().num_workers() > 1 BlockInPlace is documented in https://docs.rs/tokio/latest/tokio/task/fn.block_in_place.html

RocksDbStoreInternalError

The error type for RocksDbStoreInternal

Type Aliases

RocksDbDatabase

The RocksDbDatabase composed type with metrics

RocksDbStoreConfig

The composed config type for the RocksDbStore

RocksDbStoreError

The composed error type for the RocksDbStore