Module linera_views::batch

The definition of the batches for writing in the database. A set of functionalities for building batches to be written into the database. A batch can contain three kinds of operations on a key/value store:

Insertion of a key with an associated value

Deletion of a specific key

Deletion of all keys which contain a specified prefix

The deletion using prefixes is generally but not always faster than deleting keys one by one. The only purpose of the batch is to write some transactions into the database.

Note that normal users should not have to manipulate batches. The functionality is public because some other libraries require it. But the users using views should not have to deal with batches.

Structs

Batch

A batch of write operations.

SimpleUnorderedBatch

A batch of deletions and insertions that operate on disjoint keys, thus can be executed in any order.

SimpleUnorderedBatchIter

The iterator that corresponds to a SimpleUnorderedBatch

UnorderedBatch

An unordered batch of deletions and insertions, together with a set of key-prefixes to delete. Key-prefix deletions must happen before the insertions and the deletions.

UnorderedBatchIter

The iterator that corresponds to a SimpleUnorderedBatch

Enums

WriteOperation

A write operation as requested by a view when it needs to persist staged changes. There are 3 possibilities for the batch:

Traits

BatchValueWriter

An iterator-like object that can write values one by one to a batch while updating the total size of the batch.

DeletePrefixExpander

A trait to expand delete_prefix operations. Certain databases (e.g. DynamoDB) do not support the deletion by prefix. Thus we need to access the databases in order to replace a DeletePrefix by a vector of the keys to be removed.

LocalDeletePrefixExpander

A trait to expand delete_prefix operations. Certain databases (e.g. DynamoDB) do not support the deletion by prefix. Thus we need to access the databases in order to replace a DeletePrefix by a vector of the keys to be removed.

SimplifiedBatch

A notion of batch useful for certain computations (notably journaling).