Trait linera_views::batch::DeletePrefixExpander

pub trait DeletePrefixExpander {
    type Error: Debug;

    fn expand_delete_prefix<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key_prefix: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Vec<u8>>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

A trait to expand delete_prefix operations.

Required Associated Types

type Error: Debug

The error type that can happen when expanding the key_prefix

Required Methods

fn expand_delete_prefix<'life0, 'life1, 'async_trait>(
&'life0 self,
key_prefix: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<Vec<Vec<u8>>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Return the list of keys to be appended to the list.