Module linera_views::common

The definitions used for the KeyValueStore and Context. This provides several functionalities for the handling of data. The most important traits are:

KeyValueStore which manages the access to a database and is clonable. It has a minimal interface

Context which provides access to a database plus a base_key and some extra type E which is carried along and has no impact on the running of the system. There is also a bunch of other helper functions.

Structs

CommonStoreConfig

The common initialization parameters for the KeyValueStore

ContextFromStore

Implementation of the Context trait on top of a DB client implementing KeyValueStore.

Constants

MIN_VIEW_TAG

The minimum value for the view tags. Values in 0..MIN_VIEW_TAG are used for other purposes.

Traits

AdminKeyValueStore

Low-level trait for the administration of stores and their namespaces.

Context

The context in which a view is operated. Typically, this includes the client to connect to the database and the address of the current entry.

CustomSerialize

Sometimes we need a serialization that is different from the usual one and for example preserves order. The {to/from}_custom_bytes has to be coherent with the Borrow trait.

KeyIterable

How to iterate over the keys returned by a search query.

KeyValueIterable

How to iterate over the key-value pairs returned by a search query.

KeyValueStore

Low-level, asynchronous write and read key-value operations. Useful for storage APIs not based on views.

LocalAdminKeyValueStore

Low-level trait for the administration of stores and their namespaces.

LocalKeyValueStore

Low-level, asynchronous write and read key-value operations, without a Send bound. Useful for storage APIs not based on views.

LocalReadableKeyValueStore

Low-level, asynchronous read key-value operations. Useful for storage APIs not based on views.

LocalWritableKeyValueStore

Low-level, asynchronous write key-value operations. Useful for storage APIs not based on views.

ReadableKeyValueStore

Low-level, asynchronous read key-value operations. Useful for storage APIs not based on views.

WritableKeyValueStore

Low-level, asynchronous write key-value operations. Useful for storage APIs not based on views.

Functions

from_bytes_option

Deserializes an Optional vector of u8

get_interval

Computes an interval so that a vector has key_prefix as a prefix if and only if it belongs to the range.

get_uleb128_size

This computes the offset of the BCS serialization of a vector. The formula that should be satisfied is serialized_size(vec![v_1, …., v_n]) = get_uleb128_size(n)