Trait linera_views::common::LocalWritableKeyValueStore

pub trait LocalWritableKeyValueStore<E> {
    const MAX_VALUE_SIZE: usize;

    // Required methods
    async fn write_batch(&self, batch: Batch, base_key: &[u8]) -> Result<(), E>;
    async fn clear_journal(&self, base_key: &[u8]) -> Result<(), E>;
}

Low-level, asynchronous write key-value operations. Useful for storage APIs not based on views.

Required Associated Constants

const MAX_VALUE_SIZE: usize

The maximal size of values that can be stored.

Required Methods

async fn write_batch(&self, batch: Batch, base_key: &[u8]) -> Result<(), E>

Writes the batch in the database with base_key the base key of the entries for the journal.

async fn clear_journal(&self, base_key: &[u8]) -> Result<(), E>

Clears any journal entry that may remain. The journal is located at the base_key.

Object Safety

This trait is not object safe.