Module linera_views::scylla_db

A storage backend for views based on ScyllaDB This provides a KeyValueStore for the ScyllaDB database. The code is functional but some aspects are missing.

The current connection is done via a Session and a corresponding primary key that we name namespace. The maximum number of concurrent queries is controlled by max_concurrent_queries.

We thus implement the

KeyValueStore for a database access.

Context for the context.

Structs

ScyllaDbStore

A shared DB store for ScyllaDB implementing LruCaching

ScyllaDbStoreConfig

The type for building a new ScyllaDB Key Value Store

ScyllaDbStoreInternal

The client itself and the keeping of the count of active connections.

Enums

ScyllaDbContextError

The error type for ScyllaDbStoreInternal

Functions

create_scylla_db_common_config

Creates the common initialization for RocksDB.

create_scylla_db_test_config

Creates a ScyllaDB test store config.

create_scylla_db_test_store

Creates a ScyllaDB test store.

create_scylla_db_test_uri

Creates the URI used for the tests.

Type Aliases

ScyllaDbContext

An implementation of crate::common::Context based on ScyllaDB