Module linera_views::scylla_db

A storage backend for views based on ScyllaDB This provides a KeyValueStoreClient for the ScyllaDB database. The code is functional but some aspects are missing.

The current connection is done via a Session and a corresponding primary key that we name table_name. The maximum number of concurrent queries is controlled by max_concurrent_queries.

We thus implement the

KeyValueStoreClient for a database access.

Context for the context.

Support for ScyllaDB is experimental and is still missing important features: TODO(#935): Read several keys at once TODO(#934): Journaling operations

Structs

ScyllaDbClient

A shared DB client for ScyllaDB implementing LruCaching

ScyllaDbClientInternal

The client itself and the keeping of the count of active connections.

Enums

ScyllaDbContextError

The error type for ScyllaDbClientInternal

Constants

TEST_SCYLLA_DB_MAX_CONCURRENT_QUERIES

We limit the number of connections that can be done for tests.

TEST_SCYLLA_DB_MAX_STREAM_QUERIES

The number of connections in the stream is limited for tests.

Functions

create_scylla_db_common_config

Creates the common initialization for RocksDB

create_scylla_db_test_client

Creates a ScyllaDB test client

create_scylla_db_test_context

Creates a crate::common::Context for testing purposes.

Type Definitions

ScyllaDbContext

An implementation of crate::common::Context based on ScyllaDB