linera_views::store

Trait ReadableKeyValueStore

pub trait ReadableKeyValueStore:
    WithError
    + Send
    + Sync {
    const MAX_KEY_SIZE: usize;

    // Required methods
    fn max_stream_queries(&self) -> usize;
    fn root_key(&self) -> Result<Vec<u8>, Self::Error>;
    fn read_value_bytes(
        &self,
        key: &[u8],
    ) -> impl Future<Output = Result<Option<Vec<u8>>, Self::Error>> + Send + Sync;
    fn contains_key(
        &self,
        key: &[u8],
    ) -> impl Future<Output = Result<bool, Self::Error>> + Send + Sync;
    fn contains_keys(
        &self,
        keys: &[Vec<u8>],
    ) -> impl Future<Output = Result<Vec<bool>, Self::Error>> + Send + Sync;
    fn read_multi_values_bytes(
        &self,
        keys: &[Vec<u8>],
    ) -> impl Future<Output = Result<Vec<Option<Vec<u8>>>, Self::Error>> + Send + Sync;
    fn find_keys_by_prefix(
        &self,
        key_prefix: &[u8],
    ) -> impl Future<Output = Result<Vec<Vec<u8>>, Self::Error>> + Send + Sync;
    fn find_key_values_by_prefix(
        &self,
        key_prefix: &[u8],
    ) -> impl Future<Output = Result<Vec<(Vec<u8>, Vec<u8>)>, Self::Error>> + Send + Sync;

    // Provided methods
    fn read_value<V: DeserializeOwned>(
        &self,
        key: &[u8],
    ) -> impl Future<Output = Result<Option<V>, Self::Error>> + Send + Sync { ... }
    fn read_multi_values<V: DeserializeOwned + Send + Sync>(
        &self,
        keys: &[Vec<u8>],
    ) -> impl Future<Output = Result<Vec<Option<V>>, Self::Error>> + Send + Sync { ... }
}

Asynchronous read key-value operations.

Required Associated Constants

const MAX_KEY_SIZE: usize

The maximal size of keys that can be stored.

Required Methods

fn max_stream_queries(&self) -> usize

Retrieve the number of stream queries.

fn root_key(&self) -> Result<Vec<u8>, Self::Error>

Gets the root key of the store.

fn read_value_bytes( &self, key: &[u8], ) -> impl Future<Output = Result<Option<Vec<u8>>, Self::Error>> + Send + Sync

Retrieves a Vec<u8> from the database using the provided key.

fn contains_key( &self, key: &[u8], ) -> impl Future<Output = Result<bool, Self::Error>> + Send + Sync

Tests whether a key exists in the database

fn contains_keys( &self, keys: &[Vec<u8>], ) -> impl Future<Output = Result<Vec<bool>, Self::Error>> + Send + Sync

Tests whether a list of keys exist in the database

fn read_multi_values_bytes( &self, keys: &[Vec<u8>], ) -> impl Future<Output = Result<Vec<Option<Vec<u8>>>, Self::Error>> + Send + Sync

Retrieves multiple Vec<u8> from the database using the provided keys.

fn find_keys_by_prefix( &self, key_prefix: &[u8], ) -> impl Future<Output = Result<Vec<Vec<u8>>, Self::Error>> + Send + Sync

Finds the key matching the prefix. The prefix is not included in the returned keys.

fn find_key_values_by_prefix( &self, key_prefix: &[u8], ) -> impl Future<Output = Result<Vec<(Vec<u8>, Vec<u8>)>, Self::Error>> + Send + Sync

Finds the (key,value) pairs matching the prefix. The prefix is not included in the returned keys.

Provided Methods

fn read_value<V: DeserializeOwned>( &self, key: &[u8], ) -> impl Future<Output = Result<Option<V>, Self::Error>> + Send + Sync

Reads a single key and deserializes the result if present.

fn read_multi_values<V: DeserializeOwned + Send + Sync>( &self, keys: &[Vec<u8>], ) -> impl Future<Output = Result<Vec<Option<V>>, Self::Error>> + Send + Sync

Reads multiple keys and deserializes the results if present.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.