Module linera_views::test_utils

Helper types for tests.

Modules

test_views

Some Views that are easy to use with test cases.

Structs

NonDeterministicRng

A RNG that is non-deterministic if the platform supports it.

Functions

admin_test

Exercises the functionalities of the AdminKeyValueStore. This tests everything except the delete_all which would interact with other namespaces.

generate_random_alphanumeric_string

Get a random alphanumeric string that can be used for all tests.

generate_test_namespace

Returns a unique namespace for testing.

get_random_byte_vector

Takes a random number generator, a key_prefix and extends it by n random bytes.

get_random_key_prefix

Returns a random key_prefix used for tests

get_random_key_value_operations

A bunch of puts and some deletes.

get_random_key_values

Takes a random number generator rng, a number n and returns n random (key, value) which are all distinct with key and value being of length 8.

get_random_key_values_prefix

Takes a random number generator, a key_prefix and generates pairs (key, value) with key obtained by appending 8 bytes at random to key_prefix and value obtained by appending 8 bytes to the trivial vector. We return n such (key, value) pairs which are all distinct

get_random_kset

Builds a random k element subset of n

get_random_test_scenarios

We build a number of scenarios for testing the reads.

get_small_key_space

Appends a small value to a key making collisions likely.

make_deterministic_rng

Returns a deterministic RNG for testing.

make_nondeterministic_rng

Returns a non-deterministic RNG where supported.

random_shuffle

Shuffles the values entries randomly

run_big_write_read

DynamoDb has limits at 1M (for pagination), 4M (for write) Let us go right past them at 20M of data with writing and then reading it. And 20M is not huge by any mean. All KeyValueStore must handle that.

run_reads

This test starts with a collection of key/values being inserted into the code which is then followed by a number of reading tests. The functionalities being tested are all the reading functionalities:

run_test_batch_from_blank

Writes and then reads data under a prefix, and verifies the result.

run_writes_from_blank

Run many operations on batches always starting from a blank state.

run_writes_from_state

Run some deterministic and random batches operation and check their correctness

span_random_reordering_put_delete

A random reordering of the puts and deletes. For something like MapView it should get us the same result whatever way we are calling.

tombstone_triggering_test

That test is especially challenging for ScyllaDB. In its default settings, Scylla has a limitation to 10000 tombstones. A tombstone is an indication that the data has been deleted. That is thus a trie data structure for checking whether a requested key is deleted or not.

Type Aliases

DeterministicRng

A deterministic RNG.