linera_views

Module test_utils

Helper types for tests.

Modules

performance

Functions for computing the performance of stores.

test_views

Some Views that are easy to use with test cases.

Constants

SMALL_BYTE_UPPER_LIMIT

The size of the small value used for tests.

Functions

access_admin_test

Both checks together.

add_prefix

Adds a prefix to a list of key-values

big_read_multi_values

Reading many keys at a time could trigger an error. This needs to be tested.

exclusive_access_admin_test

A store can be in exclusive access where it stores the absence of values or in shared access where only values are stored and (key, value) once written are never modified nor erased.

get_random_byte_vector

Takes a random number generator, a key_prefix and extends it by n random bytes.

get_random_key_prefix

Returns a random key prefix used for tests

get_random_key_value_operations

A bunch of puts and some deletes.

get_random_key_values

Takes a random number generator rng, a number n and returns n random (key, value) which are all distinct with key and value being of length 8.

get_random_key_values_prefix

Takes a random number generator, a key_prefix and generates pairs (key, value) with key obtained by appending len_key random bytes to key_prefix and value obtained by creating a vector with len_value random bytes. We return n such (key, value) pairs which are all distinct.

get_random_key_values_with_sizes

Generates a list of random key-values with no duplicates

get_random_kset

Builds a random k element subset of n

get_random_test_scenarios

We build a number of scenarios for testing the reads.

namespace_admin_test

Exercises the namespace functionalities of the KeyValueDatabase. This tests everything except the delete_all which would interact with other namespaces.

root_key_admin_test

Tests listing the root keys.

run_big_write_read

DynamoDB has limits at 1 MB (for pagination), 4 MB (for write) Let us go right past them at 20 MB of data with writing and then reading it. And 20 MB is not huge by any mean. All KeyValueStore must handle that.

run_reads

This test starts with a collection of key/values being inserted into the code which is then followed by a number of reading tests. The functionalities being tested are all the reading functionalities:

run_test_batch_from_blank

Writes and then reads data under a prefix, and verifies the result.

run_writes_from_blank

Run many operations on batches always starting from a blank state.

run_writes_from_state

Run some deterministic and random batches operation and check their correctness

span_random_reordering_put_delete

A random reordering of the puts and deletes. For something like MapView it should get us the same result whatever way we are calling.

tombstone_triggering_test

That test is especially challenging for ScyllaDB. In its default settings, Scylla has a limitation to 10000 tombstones. A tombstone is an indication that the data has been deleted. That is thus a trie data structure for checking whether a requested key is deleted or not.