Module linera_views::views

Elementary data-structures implementing the views::View trait.

Modules

bucket_queue_view

The BucketQueueView implements a queue that can push on the back and delete on the front and group data in buckets.

collection_view

The CollectionView implements a map structure whose keys are ordered and the values are views.

hashable_wrapper

Wrapping a view to compute a hash.

key_value_store_view

The implementation of a key-value store view.

log_view

The LogView implements a log list that can be pushed.

map_view

The MapView implements a map with ordered keys. The MapView implements a map that can be modified.

queue_view

The QueueView implements a queue that can push on the back and delete on the front.

reentrant_collection_view

The ReentrantCollectionView implements a map structure whose keys are ordered and the values are views with concurrent access.

register_view

The RegisterView implements a register for a single value.

set_view

The SetView implements a set with ordered entries.

Enums

ViewError

Main error type for the crate.

Constants

MIN_VIEW_TAG

The minimum value for the view tags. Values in 0..MIN_VIEW_TAG are used for other purposes.

Traits

ClonableView

A ClonableView supports being shared (unsafely) by cloning it.

CryptoHashRootView

A RootView that also supports crypto hash

CryptoHashView

A View that also supports crypto hash

HashableView

A view that supports hashing its values.

Hasher

The requirement for the hasher type in HashableView.

RootView

A View whose staged modifications can be saved in storage.

View

A view gives exclusive access to read and write the data stored at an underlying address in storage.

Derive Macros

ClonableView

CryptoHashRootView

CryptoHashView

HashableView

RootView

View